Make it possible to skip output redirection and to select status format.
Introduced methods on the fluent interface to set the filename that task output is redirected to. If it is not set, output is not redirected.
Added a method to the the format that status is reported in while batch is running, e.g. flat list, grouped by state or summary.