Add option to control subscription recreation to TopicScaler.cs and uuid to batch execution containers.
When executable batches were created, they always used the names 'application' and 'input', which would make it dangerous/impossible to run two batches at the same time in the same batch account. Fixed this so it is now appends a guid to the names, making them unique per job/batch.
Added an option TopicScaler to enable recreation of subscriptions (they are still not created by default).